Update test behaviour
Remove graphfile of example1.nq since graph was set to a graphuri that already
exists in example2.nq.
Now, the test for init with mode set to "graphfiles" works like intended.
