+++ .travis.yml
@@ -46,7 +46,6 @@ script:
     - coverage run -a --source=quit tests/test_helpers.py
     - coverage run -a --source=quit tests/test_namespace.py
     - coverage run -a --source=quit tests/test_provenance.py
-    - coverage run -a --source=quit tests/test_failing_graphfiles.py
 
 after_success:
     coveralls
