--- tests/samples/applicationTests/example1.nq.graph
@@ -1,1 +1,0 @@
-http://example.org/2/
